requirements.txt
python-binance==1.0.19
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
aiohttp==3.9.1
matplotlib==3.8.2
mplfinance==0.12.10b0
schedule==1.2.1
python-dotenv==1.0.0
asyncio-throttle==1.0.2
